Moved to new testing namespace and cleaned up methods for D8.2.
