+++ tests/src/Functional/MessageUiPermissions.php
@@ -5,12 +5,12 @@
  * Definition of Drupal\message_ui\Tests\MessageUiPermissions.
  */
 
-namespace Drupal\message_ui\Tests;
+namespace Drupal\Tests\message_ui\Functional;
 
 use Drupal\Component\Render\FormattableMarkup;
+use Drupal\Tests\message\Functional\MessageTestBase;
 use Drupal\user\Entity\Role;
 use Drupal\user\RoleInterface;
-use Drupal\message\Tests\MessageTestBase;
 use Drupal\message\Entity\Message;
 use Drupal\user\UserInterface;
 
@@ -78,17 +78,18 @@ class MessageUiPermissions extends MessageTestBase {
 
     // Verify the user can't create the message.
     $this->drupalGet($create_url);
-    $this->assertResponse(403, t("The user can't create a message."));
+    // The user can't create a message.
+    $this->assertResponse(403);
 
     // Grant and check create permissions for a message.
     $this->grantMessageUiPermission('create');
     $this->drupalGet($create_url);
 
-    // If we get a valid response, create a message.
-    if ($this->assertResponse(200, t("The user can create a message."))) {
-      // Create a message at current page / url.
-      $this->drupalPostForm(NULL, array(), t('Save'));
-    }
+    // Check for valid response.
+    $this->assertResponse(200);
+
+    // Create a message.
+    $this->drupalPostForm(NULL, array(), t('Save'));
 
     // Create the message url.
     $msg_url = '/message/1';
@@ -96,20 +97,24 @@ class MessageUiPermissions extends MessageTestBase {
     // Verify the user now can see the text.
     $this->grantMessageUiPermission('view');
     $this->drupalGet($msg_url);
-    $this->assertResponse(200, "The user can view a message.");
+    // The user can view a message.
+    $this->assertResponse(200);
 
     // Verify can't edit the message.
     $this->drupalGet($msg_url . '/edit');
-    $this->assertResponse(403, "The user can't edit a message.");
+    // The user can't edit a message.
+    $this->assertResponse(403);
 
     // Grant permission to the user.
     $this->grantMessageUiPermission('edit');
     $this->drupalGet($msg_url . '/edit');
-    $this->assertResponse(200, "The user can't edit a message.");
+    // The user can't edit a message.
+    $this->assertResponse(200);
 
     // Verify the user can't delete the message.
     $this->drupalGet($msg_url . '/delete');
-    $this->assertResponse(403, "The user can't delete the message");
+    // The user can't delete the message.
+    $this->assertResponse(403);
 
     // Grant the permission to the user.
     $this->grantMessageUiPermission('delete');
@@ -117,11 +122,11 @@ class MessageUiPermissions extends MessageTestBase {
 
     // User did not have permission to the overview page - verify access
     // denied.
-    $this->assertResponse(403, t("The user can't access the over view page."));
+    $this->assertResponse(403); // The user can't access the over view page.
 
     user_role_grant_permissions($this->rid, array('administer message templates'));
     $this->drupalGet('/admin/content/messages');
-    $this->assertResponse(200, "The user can access the over view page.");
+    $this->assertResponse(200); // The user can access the over view page.
 
     // Create a new user with the bypass access permission and verify the
     // bypass.
@@ -131,7 +136,8 @@ class MessageUiPermissions extends MessageTestBase {
     // Verify the user can by pass the message access control.
     $this->drupalLogin($user);
     $this->drupalGet($create_url);
-    $this->assertResponse(200, 'The user can bypass the message access control.');
+    // The user can bypass the message access control.
+    $this->assertResponse(200);
   }
 
   /**
